backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.8.0
pydantic-settings==2.1.0
email-validator==2.1.0.post1
